python-workers/requirements.txt
# Embedding Worker
bge-m3 嵌入模型 - 文本向量化

# FAISS
faiss-cpu>=1.7.4

# ONNX Runtime
onnxruntime>=1.16.0

# Transformers
transformers>=4.36.0
torch>=2.0.0

# 文件解析
PyMuPDF>=1.23.0  # PDF
python-docx>=1.1.0  # Word
openpyxl>=3.1.0  # Excel
beautifulsoup4>=4.12.0  # HTML
lxml>=5.0.0  # XML/HTML 解析

# 工具
numpy>=1.24.0
chardet>=5.2.0  # 编码检测
